// src/modules/classifications/application/dtos/classification.dto.ts
import { z } from 'zod'

export const classificationDTOSchema = z.object({
  id: z.string(),
  name: z.string(),
  description: z.string(),
  categoryType: z.string(),
})

export type ClassificationDTO = z.infer<typeof classificationDTOSchema>

export const classificationFullDTOSchema = classificationDTOSchema.extend({
  createdAt: z.string(),
})

export type ClassificationFullDTO = z.infer<typeof classificationFullDTOSchema>